pyproject.toml
[project]
name = "aiops-exp"
version = "0.1.0"
description = "A lightweight AIOps experiment runner"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12.0",
    "PyYAML>=6.0",
]

[project.scripts]
aiops-bench = "aiops_bench.cli:main"

[tool.uv]
package = true

[tool.setuptools.packages.find]
include = ["aiops_bench*"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
